third_party/WebKit/LayoutTests/platform/win/external/wpt/html/browsers/browsing-the-web/history-traversal/persisted-user-state-restoration/scroll-restoration-basic-expected.txt

This is a testharness.js-based test.
PASS Default value is "auto" 
PASS It is writable 
PASS Invalid values are ignored 
Harness: the test ran to completion.
